modificado service

+++ pratica-4-persistencia-dados-react-native/src/services/LivroService.ts
@@ -3,6 +3,7 @@ import { EntityRepository } from "../repository/base";
 import { TPayloadCreateLivro, TPayloadUpdateLivro } from "../states/livro/reducer/types";
 
 export class LivroService {
+
     private _repository: EntityRepository<Livro>;
 
     constructor() {
@@ -16,6 +17,10 @@ export class LivroService {
         livro.genero = item.genero;
 
         const result = this._repository.create(livro);
+        
+        if(livro == null) {
+            return null;
+        }
 
         return result;
     }
@@ -24,7 +29,7 @@ export class LivroService {
         const livro = await this._repository.findById(item.id);
         
         if(livro == null) {
-            throw new Error('N')
+            return null;
         }
 
         livro.titulo = item.titulo;
@@ -38,6 +43,10 @@ export class LivroService {
 
     async delete(id: number) {
         const livro = await this.findById(id);
+        
+        if(livro == null) {
+            return null;
+        }
 
         this._repository.delete(livro);
     }
@@ -46,12 +55,14 @@ export class LivroService {
         const livro = await this._repository.findById(id);
         
         if(livro == null) {
-            throw new Error('N')
+            return null;
         }
 
         return livro;
     }
 
     async findAll() {
+        const livros = await this._repository.findAll();
+        return livros;
     }
 }
